requirements.txt
langchain
langgraph
langchain-groq
ipykernel
python-dotenv
openai
pyttsx3
gTTS
pyaudio
speechrecognition
sounddevice
numpy
scipy
pygame
Pyside6
PyQt6
pydub
edge_tts
soundfile
